+++ crates/apollo-leto-interop/src/lib.rs
@@ -8,7 +8,7 @@ mod array;
 mod view;
 
 pub use array::{
-    try_array1_from_slice, try_array1_from_vec, try_dense_from_array, try_dense_from_contiguous,
-    try_dense_from_slice, try_dense_from_view,
+    try_array1_from_slice, try_array1_from_vec, try_dense_from_array, try_dense_from_slice,
+    try_dense_from_view,
 };
 pub use view::view_cow;